requirements.txt
pyowm
netifaces
python-daemon
